Organized run report code
