+++ src/helpers/RunReportHelper.ts
@@ -1,0 +1,58 @@
+import { Repositories } from "../repositories";
+import { ArrayHelper } from "../apiBase";
+import { Parameter, Query, Report, ReportResult } from "../models";
+
+export class RunReportHelper {
+
+  public static async runAllQueries(report: Report) {
+    let depth = 0;
+    let keepGoing = true;
+    while (keepGoing) {
+      keepGoing = await RunReportHelper.runQueries(report, depth);
+      if (keepGoing) RunReportHelper.populateQueryResultParameters(report, depth)
+      depth++;
+    }
+  }
+
+  private static populateQueryResultParameters(report: Report, depth: number) {
+    report.parameters?.forEach(p => {
+      if (p.source === "query") {
+        const table = p.sourceKey.split('.')[0];
+        const field = p.sourceKey.split('.')[1];
+        report.queries.forEach(q => {
+          if (q.keyName === table && q.depth === depth && q.value) {
+            p.value = [];
+            q.value.forEach(v => { p.value.push(v[field]); })
+          }
+        })
+      }
+    })
+  }
+
+  private static async runQueries(report: Report, depth: number) {
+    const queries: Query[] = ArrayHelper.getAll(report.queries, "depth", depth);
+    if (queries.length > 0) {
+      const promises: Promise<void>[] = []
+      queries.forEach(q => { promises.push(this.runQuery(q, report)) });
+      await Promise.all(promises)
+      return true;
+    } else return false;
+  }
+
+  private static async runQuery(query: Query, report: Report) {
+    const parameters: any[] = [];
+
+    const storedSql = query.sqlLines.join(" ");
+
+    storedSql.match(/:[A-Za-z0-9]{1,99}/g)?.forEach(m => {
+      const keyName = m.replace(":", "");
+      const p: Parameter = ArrayHelper.getOne(report.parameters, "keyName", keyName)
+      parameters.push(p.value);
+    });
+
+    let sql = storedSql;
+    storedSql.match(/:[A-Za-z0-9]{1,99}/g)?.forEach(m => { sql = sql.replace(m, "?") });
+    query.value = await Repositories.getCurrent().report.run(query.db, sql, parameters);
+  }
+
+}
