+++ src/helpers/index.ts
@@ -4,3 +4,5 @@ export * from "./MembershipPermissions";
 export * from "./ReportingPermissions";
 export * from "./AttendancePermissions";
 export * from "./DonationPermissions";
+export { ReportResultHelper } from "./ReportResultHelper";
+export { RunReportHelper } from "./RunReportHelper";
